E2Eテストのログインを共通化
